Correct an addition made recently (Sept, r5550 & r5551) to use the current directory for a sub-script instead of the full path.  This is needed when running 'devtest' outside of the tinderbox environment.
In the context in which this is run, the current directory should be
the one that contains the program being run and so the full path isn't
needed and is incorrect except when running in the tinderbox
continuous test environment at DICE-UCSD.

I noticed this in testing on a non-tinderbox host yesterday.  This
would have also been noticed in the upcoming pre-release NMI BatLab
(Build and Test) runs.



git-svn-id: svn://irodssvn.ucsd.edu/trunk@5617 ea137bd6-aa86-47b6-b90d-b101d9f19a90