Set the -DSHA256_FILE_HASH switch in platform.mk[.template] so that all source will be built with it.
This is adding to the recent updates to support SHA2 file hashes.

Since the dataObjInfo_t structure is changed for the new CHKSUM_LEN
size, and that's used in many functions throughout, we need to be sure
that the -DSHA256_FILE_HASH is added for all.  So we can just set it
in the platform.mk.  Without this, we were still seeing mismatches
(and fatal errors) in functions used by isysmeta.

With this addition, with SHA256_FILE_HASH enabled, 'irodsctl devtest'
now succeeds.


git-svn-id: svn://irodssvn.ucsd.edu/trunk@5621 ea137bd6-aa86-47b6-b90d-b101d9f19a90