Add code to clientLoginWithPassword, matching that in clientLogin, to save a signiture of the session challenge for use in changing another's password.
This signiture is needed for the administrative function of changing
another user's password.  The problem is not seen in iadmin since it
uses clientLogin.

The clientLoginWithPassword code, including this change, is exercised
via one of the ICAT tests under 'devtest' but only to confirm that it
can successfully log in.  To verify that this change works, I made a
temporary change to iadmin.c to call clientLoginWithPassword instead
of clientLoging and I verified that without this fix it does fail
(gets a CAT_PASSWORD_ENCODING_ERROR error) and with this fix it seems
to work fine.

This was reported in bugzilla item 167 earlier today.



git-svn-id: svn://irodssvn.ucsd.edu/trunk@5626 ea137bd6-aa86-47b6-b90d-b101d9f19a90