Update iadmin help text to explain that a remote zone definition needs to specify the IES, not a non-IES.
When a non-ICAT-Enabled-Server is used, the error returned to remote
zone users trying to connect is CAT_INVALID_USER (even if they have a
valid local user account defined).  This is inaccurate and misleading
but it would be difficult to do better (to determine if this is caused
by a non-IES remote zone definition).  What happens to the server is
non-fatal, but prevents proper authentication.  Hopefully, this
warning in the iadmin help text for mkzone and modzone will be
sufficient.



git-svn-id: svn://irodssvn.ucsd.edu/trunk@5639 ea137bd6-aa86-47b6-b90d-b101d9f19a90