Improve the error message associated with CAT_PASSWORD_ENCODING_ERROR, in particular explaining that it can fail with the admin is using GSI.
The problem is that the 'iadmin moduser Name password pw' command uses
the password as a shared secret in obfuscating the entered password
and that's not available when GSI or Kerberos is being used.  

This can also happen if there's no .irodsA file (the admin enters the
password to go into iadmin interactive mode), and then runs the mod
password command, is prompted for their password again, and enters it
incorrectly.

The problem with this with GSI was mentioned to me by Jean-Yves
recently.  I cannot easily fix this, but this error message should
help, providing clues on how to work around the problem.



git-svn-id: svn://irodssvn.ucsd.edu/trunk@5660 ea137bd6-aa86-47b6-b90d-b101d9f19a90