Addition for previous commit; need to set some attriInx values in the general-query result structure for it to work properly in all cases.
The problem was noticed in additional testing at iPlant and via
'devtest' at DICE-UCSD (with the specific-query defined) but this
seems to correct the issues now.

In other uses of the specific-query the results are returned to the
client and handled fine, but with these changes we can now also use a
specific-query to replace an internally used general-query.



git-svn-id: svn://irodssvn.ucsd.edu/trunk@5683 ea137bd6-aa86-47b6-b90d-b101d9f19a90