This is a continuation of the previous commits (April 3 and 4) to implement a specific-query to improve performance at scale for certain server operations using new SQL from iPlant.  Tony Edgin of iPlant did additional testing and found this problem when large numbers for files exist in a sub-collections.
Without this fix, a 'imv' of a collection with subcollections with
files exceeding the MAX_SQL_ROWS (256) will get an error
(UNMATCHED_KEY_OR_INDEX) and, at least in some cases, failover to the
general-query (rsGenQuery) (which works, but can be slow).

The fix is just the one line below, which is possible because of the
layering of the ICAT code; the mid-level ICAT functions can return
rows thru the general-query interface even tho the original query was
via the specific-query.



git-svn-id: svn://irodssvn.ucsd.edu/trunk@5684 ea137bd6-aa86-47b6-b90d-b101d9f19a90