This a new i-command, 'ifind', similar to the Unix/Linx 'find' utility, that can be a useful tool for searching iRODS namespace.  See the help text (in the 'usage' function) for a fairly extensive description.
This is intended for use with iRODS 3.3.1 and is also planned to be
integrated with the 4.x line in the near future.  It is freely
available to the community.  To add it to a 3.3.1 instance, add these
mods to your source tree (the new ifind.c and Makefile updates) and
run 'make'.

This development was supported by NSF grant #1032468 "SDCI Sec
Improvement: Building from Bedrock: Infrastructure Improvements for
Collaboration and Science" via a subcontract to DICE-UCSD.

The above subcontract supported the development of 'ifind' and two
other capabilities: a partial copy system (implemented as an 'ipc'
i-command), and some extensions for providing environment variables on
the command line.



git-svn-id: svn://irodssvn.ucsd.edu/trunk@5685 ea137bd6-aa86-47b6-b90d-b101d9f19a90