This a new i-command, 'ipc', to perform partial copies of files to/from iRODS space.  See the help text (in the 'usage' function) for a description.
We had considered modifying iget and iput and supporting functions for
this, but determined that that would be overly complicated and risky
(could easily break some of the existing logic), particularly when
trying to integrate this into both 3.3.1 and 4.x.  Implemented as a
stand-alone i-command, was straight-forward and we expect will be
easier for users to use.

This is intended for use with iRODS 3.3.1 and is also planned to be
integrated with the 4.x line in the near future.  It is freely
available to the community.  To add it to a 3.3.1 instance, add these
mods to your source tree (the new ipc.c and Makefile updates) and
run 'make'.

This development was supported by NSF grant #1032468 "SDCI Sec
Improvement: Building from Bedrock: Infrastructure Improvements for
Collaboration and Science" via a subcontract to DICE-UCSD.

The above subcontract supported the development of 'ipc' and two other
capabilities: an iRODS 'find' i-command ('ifind') and some extensions
for providing environment variables on the command line.



git-svn-id: svn://irodssvn.ucsd.edu/trunk@5686 ea137bd6-aa86-47b6-b90d-b101d9f19a90