These updates allow a set of iRODS environment variables to be provided on the command line.  The new command line options are: --irodsHost --irodsPort --irodsUserName --irodsZone --irodsHome --irodsCwd --irodsDefResource
If the first four are included, no .irodsEnv file is needed.  
See the revised help text in ienv.c (in the 'usage' function) for a
description.

The i-commands call parseCmdLineOpt and for these new env options to
be avaiable that must be called with the the long options enabled
(optStr argument needs to include 'Z' and includeLong set).
i-commands for which this is true are: ibun, ichksum, ienv, iexecmd,
iget, iinit, ils, imv, incarch, incattr, inc, iput, iqstat, iquest,
ireg, irepl, irm, irmtrash, irsync, irule, and itrim. To add others,
modify the optStr (an input to parseCmdLineOpt) to include Z and
change the input flag from 0 to 1 (the includeLong flag), like was
done for ienv.c (see diff's included in this commit).

This is intended for use with iRODS 3.3.1 and is also planned to be
integrated with the 4.x line in the near future.  It is freely
available to the community.  To add it to a 3.3.1 instance, add these
mods to your source tree and run 'make'.

This development was supported by NSF grant #1032468 "SDCI Sec
Improvement: Building from Bedrock: Infrastructure Improvements for
Collaboration and Science" via a subcontract to DICE-UCSD.

The above subcontract supported this development and two other
capabilities: an iRODS 'find' i-command ('ifind') and a partial copy
system (implemented as an 'ipc' i-command).



git-svn-id: svn://irodssvn.ucsd.edu/trunk@5687 ea137bd6-aa86-47b6-b90d-b101d9f19a90