Add the new 'ifind' and 'ipc' i-commands to 'ihelp'.
This commit updates the 'ihelp' command to include the new 'ifind' and
'ipc' i-commands in its summary list and for running them (like
others) to provide more detailed information.  This is a continuation
of commits made on August 28, where 'ifind', 'ipc', and command-line
environment variables were added.

This is intended for use with iRODS 3.3.1, is also planned to be
integrated with the 4.x line in the near future, and is freely
available to the community.

This development was supported by NSF grant #1032468 "SDCI Sec
Improvement: Building from Bedrock: Infrastructure Improvements for
Collaboration and Science" via a subcontract to DICE-UCSD.



git-svn-id: svn://irodssvn.ucsd.edu/trunk@5688 ea137bd6-aa86-47b6-b90d-b101d9f19a90