This is the patch we issued on December 3rd via irod-chat and the wiki (iRODS_3.3.1.Patch_1 and Downloads pages) for some security vulnerabilities.  Hao developed these fixes and then backported them to 3.3.1.
Keith James asked us today via github to commit this to svn, so I
went ahead and did so.

Note that the 3.3.1 release tar file still does not have this patch,
as that's how we have done it in the past.  We don't update the
release, but list source patches that should be applied.

This is the only patch we have for 3.3.1.

 - Wayne -



git-svn-id: svn://irodssvn.ucsd.edu/trunk@5689 ea137bd6-aa86-47b6-b90d-b101d9f19a90